requirements.txt
geopandas
pandas
shapely
rasterio
pyproj
requests
torch
torchvision
Pillow
openpyxl
folium
gitpython
seaborn
numpy
scikit-learn
ultralytics
opencv-python-headless